scripts/train.py
#!/usr/bin/env python3
"""
MOUAADNET-ULTRA: Training Script
=================================
Production training script with full configuration support.

Usage:
    python scripts/train.py --config configs/default.yaml
    python scripts/train.py --data /path/to/data --epochs 100
"""

import argparse
import sys
from pathlib import Path

# Add project root to path
sys.path.insert(0, str(Path(__file__).parent.parent))

import yaml
import torch
from mouaadnet_ultra.model import MouaadNetUltra, MouaadNetUltraLite, MouaadNetUltraPro


def parse_args():
    parser = argparse.ArgumentParser(
        description="Train MOUAADNET-ULTRA",
        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
    )
    
    # Config
    parser.add_argument("--config", type=str, default="configs/default.yaml",
                        help="Path to config file")
    
    # Data overrides
    parser.add_argument("--data", type=str, help="Path to dataset")
    parser.add_argument("--img-size", type=int, help="Input image size")
    parser.add_argument("--batch-size", type=int, help="Batch size")
    
    # Training overrides
    parser.add_argument("--epochs", type=int, help="Number of epochs")
    parser.add_argument("--lr", type=float, help="Learning rate")
    parser.add_argument("--resume", type=str, help="Resume from checkpoint")
    
    # Model
    parser.add_argument("--model", type=str, choices=["ultra", "lite", "pro"],
                        default="ultra", help="Model variant")
    
    # Hardware
    parser.add_argument("--device", type=str, default="cuda", help="Device")
    parser.add_argument("--workers", type=int, default=8, help="DataLoader workers")
    
    # Output
    parser.add_argument("--save-dir", type=str, default="checkpoints",
                        help="Save directory")
    parser.add_argument("--name", type=str, default="exp", help="Experiment name")
    
    return parser.parse_args()


def load_config(config_path: str) -> dict:
    """Load YAML configuration."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def create_model(variant: str, config: dict):
    """Create model based on variant."""
    model_classes = {
        "ultra": MouaadNetUltra,
        "lite": MouaadNetUltraLite,
        "pro": MouaadNetUltraPro,
    }
    
    model_class = model_classes.get(variant, MouaadNetUltra)
    return model_class()


def main():
    args = parse_args()
    
    print("=" * 60)
    print("MOUAADNET-ULTRA Training")
    print("=" * 60)
    
    # Load config
    config = {}
    if Path(args.config).exists():
        config = load_config(args.config)
        print(f"Loaded config: {args.config}")
    
    # Apply CLI overrides
    if args.epochs:
        config.setdefault("training", {})["epochs"] = args.epochs
    if args.lr:
        config.setdefault("training", {})["learning_rate"] = args.lr
    if args.batch_size:
        config.setdefault("data", {})["batch_size"] = args.batch_size
    
    # Create model
    model = create_model(args.model, config)
    print(f"\nModel: {args.model.upper()}")
    print(f"Parameters: {model.count_parameters():,}")
    print(f"Size: {model.get_model_size_mb():.2f} MB")
    
    # Device
    device = torch.device(args.device if torch.cuda.is_available() else "cpu")
    print(f"Device: {device}")
    
    # TODO: Add actual training loop
    # For now, this is a placeholder showing the structure
    print("\n⚠️  Note: Currently using placeholder training.")
    print("    Integrate with your dataset for actual training.")
    
    print("\n✓ Training script ready!")
    print("=" * 60)


if __name__ == "__main__":
    main()
